Separate APIs of each class into individual files
