+++ backend/my_queue/api/__init__.py
@@ -1,0 +1,11 @@
+from .business import BusinessController
+from .queue import QueueController
+from .entry import EntryController
+from .analytic import AnalyticController
+
+__all__ = [
+    AnalyticController,
+    BusinessController,
+    EntryController,
+    QueueController,
+]
